Add files via upload
It is all about developing a mobile application, which can be fueled by AI to provide people in need with the opportunity to get emotional support on a personal scale at an affordable cost. Anxiety, stress and depression are becoming more common in the whole world, which are issues that affect mental health. However, the difficulty lies in seeking professional assistance because it is expensive, stigmatized by the society, and there are no qualified specialists. This is where the project comes in, it will be trying to bridge that gap by offering the users with a constant, reliable support that is only a tap away, wherever and whenever that person needs it. Some of the key functions in the app include the therapy procedures, mood tracker, along with the existing methods, including the cognitive behavioral therapy (CBT). All these features are very well thought-out to allow users to keep track of their mental health, identify negative thoughts, and shift to more productive coping mechanisms. Moreover, the app guarantees privacy and ethical issues to the users by ensuring the data handling practices are secure and, therefore, give users the confidence to address sensitive mental health problems. By far the most important innovation that falls within the project scope is the use of conversational artificial intelligence that will facilitate the use of a chatbot with the ability to interpret user messages in real-time to create useful and enjoyable experiences. Besides that, sentiment analysis lands the system on facial triggers and retreats personalized, empathetic messages. 
It is true that this project illustrates the potential of the artificial intelligence technologies with the help of the traditional mental care to lower access barriers and encourage preventive care of mental health. The combination of innovative technologies and moral responsibility is expected to help this mobile application to offer the scalable, universal, and meaningful solution to the emerging demand of the mental care that is proliferating in our virtual world.