Update README.md
This Project work was Submitted to the Department of Information Technology and Decision Sciences, University of Energy and Natural Resources, Sunyani in Partial Fulfillment of the Requirements for Degree in INFORMATION TECHNOLOGY
SEPTEMBER, 2025.
Along the line we faced so many errors and we started many times with different project names. The last time we figured it out, the project was named "start". Becasue that was the name we could think of starting at that point and changing the name later on might have caused dependancy issues so we left it like that and everything works perfectly.